app/requirements.txt
fastapi
uvicorn
sentence-transformers
torch
python-dotenv
